third_party/blink/web_tests/fast/dom/Element/hostname-host-expected.txt

http with default port
https with default port
ftp with default port

http with non-default port
https with non-default port
ftp with non-default port

http without port
https without port
ftp without port

http with port 0
Hostname is dev.w3.org. Host is dev.w3.org
Hostname is dev.w3.org. Host is dev.w3.org
Hostname is dev.w3.org. Host is dev.w3.org
Hostname is dev.w3.org. Host is dev.w3.org:123
Hostname is dev.w3.org. Host is dev.w3.org:123
Hostname is dev.w3.org. Host is dev.w3.org:123
Hostname is dev.w3.org. Host is dev.w3.org
Hostname is dev.w3.org. Host is dev.w3.org
Hostname is dev.w3.org. Host is dev.w3.org
Hostname is dev.w3.org. Host is dev.w3.org:0
